requirements.txt
pytest==7.2.2
Django~=3.2.18
djangorestframework~=3.14.0
eventbrite~=3.3.5